Move versions to their own file, add tests
